refactor: separate remote backend implementations from mixin
