+++ docarray/array/array/pushpull.py
@@ -23,13 +23,7 @@ SUPPORTED_PUSH_PULL_PROTOCOLS = get_args(PUSH_PULL_PROTOCOL)
 
 if TYPE_CHECKING:  # pragma: no cover
     from docarray import BaseDocument, DocumentArray
-    from docarray.array.array.pushpull.abstract_doc_store import AbstractDocStore
-
-
-class ConcurrentPushException(Exception):
-    """Exception raised when a concurrent push is detected."""
-
-    pass
+    from docarray.remote.abstract_doc_store import AbstractDocStore
 
 
 SelfPushPullMixin = TypeVar('SelfPushPullMixin', bound='PushPullMixin')
@@ -69,17 +63,17 @@ class PushPullMixin(Iterable['BaseDocument']):
             return cls.__backends__[protocol]
 
         if protocol == 'jinaai':
-            from docarray.array.array.pushpull.jinaai import JACDocStore
+            from docarray.remote.jinaai import JACDocStore
 
             cls.__backends__[protocol] = JACDocStore
             logging.debug('Loaded Jina AI Cloud backend')
         elif protocol == 'file':
-            from docarray.array.array.pushpull.file import FileDocStore
+            from docarray.remote.file import FileDocStore
 
             cls.__backends__[protocol] = FileDocStore
             logging.debug('Loaded Local Filesystem backend')
         elif protocol == 's3':
-            from docarray.array.array.pushpull.s3 import S3DocStore
+            from docarray.remote.s3 import S3DocStore
 
             cls.__backends__[protocol] = S3DocStore
             logging.debug('Loaded S3 backend')
