--- docarray/array/array/pushpull/__init__.py
@@ -1,3 +1,0 @@
-from docarray.array.array.pushpull.pushpull import PushPullMixin
-
-__all__ = ['PushPullMixin']
