+++ docarray/remote/exceptions.py
@@ -1,0 +1,4 @@
+class ConcurrentPushException(Exception):
+    """Exception raised when a concurrent push is detected."""
+
+    pass
