+++ tests/integrations/remotes/test_hubble.py
@@ -9,7 +9,7 @@ from docarray.documents import TextDoc
 from tests.integrations.remotes import gen_text_docs, get_test_da, profile_memory
 
 DA_LEN: int = 2**10
-TOLERANCE_RATIO = 0.1  # Percentage of difference allowed in stream vs non-stream test
+TOLERANCE_RATIO = 0.5  # Percentage of difference allowed in stream vs non-stream test
 RANDOM: str = uuid.uuid4().hex[:8]
